Restructured the test class.
